+++ sdk/highlight-remix/src/highlight-init.tsx
@@ -1,6 +1,8 @@
 import { H } from 'highlight.run'
 import type { HighlightOptions } from 'highlight.run'
 import React, { useEffect } from 'react'
+import { SESSION_STORAGE_KEYS } from '@highlight-run/client/src/utils/sessionStorage/sessionStorageKeys'
+import Cookies from 'js-cookie'
 
 interface Props extends HighlightOptions {
 	projectId?: string
@@ -8,7 +10,18 @@ interface Props extends HighlightOptions {
 
 export function HighlightInit({ projectId, ...highlightOptions }: Props) {
 	useEffect(() => {
-		projectId && H.init(projectId, highlightOptions)
+		if (projectId) {
+			const { sessionSecureID } = H.init(projectId, highlightOptions) || {
+				sessionSecureID: '',
+			}
+
+			if (sessionSecureID) {
+				Cookies.set(
+					SESSION_STORAGE_KEYS.SESSION_SECURE_ID,
+					sessionSecureID,
+				)
+			}
+		}
 	}, []) // eslint-disable-line react-hooks/exhaustive-deps
 
 	return null